Fixed combo box parameters displayed and their order, fixed BEs to be able to search for Date
