+++ Part_2/airlinesystem/Passenger.java
@@ -1,7 +1,9 @@
 package airlinesystem;
 import java.io.*;
 import java.net.Socket;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 
 public class Passenger {
     protected PrintWriter ticketWriter;
@@ -139,6 +141,23 @@ public class Passenger {
 		    }
 		    return true;
 	}
+    
+    public boolean isValidDate(String date){
+		//code obtained and edited from http://stackoverflow.com/questions/15491894/regex-to-validate-date-format-dd-mm-yyyy
+		if(!date.matches("^(?=\\d{2}([/])\\d{2}\\1\\d{4}$)(?:0[1-9]|1\\d|[2][0-8]|29(?!.02.(?!(?!(?:[02468][1-35-79]|[13579]"
+				+ "[0-13-57-9])00)\\d{2}(?:[02468][048]|[13579][26])))|30(?!.02)|31(?=.(?:0[13578]|10|12))).(?:0[1-9]|1[012]).\\d{4}$"))
+			return false;
+		SimpleDateFormat format = new SimpleDateFormat("dd/MM/yyyy");
+		Date currententDate = new Date();
+		try{
+			Date flightDate = format.parse(date);
+			if(!flightDate.after(currententDate))
+				return false;
+		}catch (Exception e){
+			return false;
+		}
+		return true;
+	}
 
     public String checkFormatFlightSearch(String param, String key){
     	String toReturn = "";
@@ -154,6 +173,10 @@ public class Passenger {
     	case "source":
     		if(key.length() > 45 || key.contains("_"))
     			toReturn = "Search field format error, please ensure field does not contain underscores or exceed 45 characters";
+    		break;
+    	case "date":
+    		if(!isValidDate(key))
+    			toReturn = "Search field format error, please ensure field follows dd/MM/yyyy format";
     		break;
     	default:
     		toReturn = "Unrecognized Parameter error...";
